41753: Lint API Check: Consult source files for constants
Lint currently checks the .class files for field references and method
references to APIs that require a version higher than the
minSdkVersion in the manifest.

However, constants (such as final static integers) will be copied into
the .class file, so there is no reference to the original API and its
API version, which means lint can't flag these.

In some cases, such as referencing LayoutParams.MATCH_PARENT, that's
no problem; the constant value works on older versions, and there is
no problem, since the value rather than the reference is
used. However, in other cases this may lead to runtime crashes.

This CL updates lint to look at the source files and flag field
references.  It excludes some constants that are known to be okay
(such as referencing say android.os.Build.VERSION_CODES.JELLY_BEAN_MR1
(which is typically done precisely to conditionally branch based on
the current device's version). We may need to do additional tweaks to
this in the future, since unlike method and field references there are
probably many cases where referencing these constants are fine on
older devices, but there are also cases where it's dangerous. For now,
we're erring on the side of warning the developer rather than being
certain that it's not a problem.

Change-Id: Ia361a5a9143a4662c21b9d987ea7afba84b4519e