Don't cache device name obtained when device is offline
When a device is connected, it goes from offline to online status
in a short time. However, the name for the device (obtained by
reading device's system properties) is cached the first time it is
requested. This defaults to just the serial number since the
device properties can't be read when the device is offline. This CL
just makes sure that the device name is cached only when the device
is online.

Change-Id: Iff11f46ed9e8b25508b72cc002cd57741f7b6c8d