Couple of internal plugins for Gradle.
'distrib' allows copying output jars and their dependencies
in a distribution folder in a way that will match the tools
component of the SDK. It will also copy the launcher scripts,
and handle NOTICE files.

'clone-artifacts' is a plugin to pull a clone of some maven
artifacts to allow running Gradle in offline mode.

Change-Id: I9a3321d17c2194e28bc2edefbe74087093812884