Update IShellEnabledDevice to only use the new method.
Since we added the method with the TimeUnits, the other one
is now deprecated. However the extraction of this method in
an interface is fairly new, so I'm moving the deprecated
version back in IDevice.

This makes IShellEnabledDevice cleaner, and since this is a published
API in the Gradle build system (as part of the DeviceConnector API for
remote device access), this is a good thing.

Also trying to fix the RemoteAndroidTestRunnerTest...

Change-Id: Iee5a1def46858086fd7d73881890cad02b85cae8