Add bcc_compat to Ant.
Refactor the renderscript compilation to only run llvm
once to compile all the .rs files. This requires significant changes
to how we detect that recompilation is needed, so the old dependency
graph is not used anymore.

A new processor class will allow reusability with ADT and possibly
Gradle.

Also changed BuildToolsInfo to add support for bcc_compat, adding
support for new tools based on a min revision so that validity testing
is not impacted for older versions.

Change-Id: I724ba57fab3abe263e08472817d8fd2160e27362