9patch viewer: Fit in with current L&F
The status panel at the bottom of the 9patch viewer was customized
to be black with the labels set in white. This doesn't work when
using custom L&F such as under IDEA's Alloy or Darcula themes.

This CL:
 - Removes the customization of foreground colors for labels.
   They now inherit from the default L&F.
 - Makes the main panel of the image viewer to be a lighter
   variation of the panel background color from the current L&F.

This allows it to work standalone and fit in with the IDE's current
L&F.

Fixes: https://code.google.com/p/android/issues/detail?id=58254

Change-Id: I57d0aeb52a89795c51fd9736b8d8555ad15e6bfe