Do not customize label colors.
The status panel at the bottom was customized to be of a dark color
with the labels set in white. This doesn't work when using custom
L&F such as under IDEA's Alloy or Darcula themes. This CL removes
the styling since the defaults work well under all L&F's.

Change-Id: I57d0aeb52a89795c51fd9736b8d8555ad15e6bfe