Add plus addressing to MavenCoordinate
This change adds support for using a single '+' as a revision
for a gradle coordinate.
It also makes the addressing more flexible, allowing for
arbitrary numbers of parts to a revision number.
Since this functionality is only present in Gradle, not Maven in
general, this class has been renamed to GradleCoordinate.
Change-Id: I4d4a33c8f4ebe2caf8396af0a4d2e9516892de55