Parse system property for ANDROID_ADB_SERVER_PORT.
This commit adds the support for taking system property
ANDROID_ADB_SERVER_PORT into consideration when parsing port for adb
server. Before this, only environment variable was taken into account so
it was not possible to setup port in runtime programmatically since
environment property is set upon JVM startup and can not be changed
anymore.

When parsing, system property is taken into account firstly. When not
set, env property is parsed as earlier. There is preserved backward
compatibility with previous versions.

https://code.google.com/p/android/issues/detail?id=62645

Change-Id: I16162e91dd013f9ac76ebe8276748e0c4dfd175e
Signed-off-by: Stefan Miklosovic <miklosovic@gmail.com>