Increase severity from error to fatal for a few lint issues
The distinction between a "fatal" error and just an "error" is that fatal issues
are checked when you try to export to APK in Eclipse, and also while building
a release target from Gradle (this is added in a separate CL).

This CL increases the severity for a number of issues from error to fatal such
that they are included in this release check.

Change-Id: Ic8291cdffcec859c93528be7480803ec15d747d3