Load dx.jar inside the plugin.
This is to prevent calling out to the command line version
which will spawn a new JVM. In large projects with lots of
modules or lots of dependencies, this can happen many, many
times (due to threaded pre-dex).

The Wrapper is coming from the one in ADT. However it is not
reentrant as its usage depends on static things inside dx.

A pool of dex wrapper allows creating more than one wrapper
(loading dx.jar multiple times), and reuse them when they are
no longer needed.

Change-Id: Ie31145f6ca3d37f35e7b0f7005f99608e257fa2b