WIP: lintCompile task
Adds a compileLintRules task which compiles sources found
in src/lint/ using the lint-api.jar and dependent classes
as the classpath for the custom rules. These are then packaged
into lint.jar, which in turn is picked up by the library's
AAR assembly task to package lint.jar, and the lint task
to be included as lint rules on the project check.

Change-Id: I4c34093f1a1fd0b4b7e5bfa1015486e9f275d6f6