Attempt to solve deadlock issues in EmulatorConsole.
EmulatorConsole#readLines can block indefinitely if emulator is
unresponsive.

It was previously called within a static synchronized block, which
would cause all other attempts to get a EmulatorConsole,
even for different emulators, to block.

The problem the socket channel used by EmulatorConsole was in
blocking mode, and the read and write algorithms expect a non
blocking socket channel.

This commit also changes the EmulatorConsole#getConsole method to not
block on the class while a connection to one emulator console is being
attempted. Instead, a lock on just the static hash map sEmulators
is obtained when reading or writing to that hashmap. And rely
on object level locks on EmulatorConsole to synchronize between
threads that are trying to use EmulatorConsole for the same
emulator.

Bug: 13061931
Change-Id: I427a94ca23105ba47eec9c0df01c903af7bdd3d0