Attempt to solve deadlock issues in EmulatorConsole.
EmulatorConsole#readLines can go into a seemingly infinite loop.
It was previously called within a static synchronizd block, which
would cause all other attempt to get a EmulatorConsole to block.

This commit relaxes the EmulatorConsole#readLines exit criteria,
to ~5 seconds of attempts to read rather than just 5 seconds of
failed attempts. It also changes the EmulatorConsole#getConsole
method to not block on the class while a connection to one emulator
console is being attempted.

Bug: 13061931
Change-Id: I427a94ca23105ba47eec9c0df01c903af7bdd3d0