More incremental lint checks
While many lint checks can run incrementally in the editor, such
as the API check, many other checks can not -- because they depend on
looking at multiple sources of data. Take the String.format checker
for example; it needs to check that the types of the parameters passed
to the formatting call (such as 42) match the formatting strings found
in the XML resources (such as %d).

This changeset allows lint clients such as Studio to provide direct
lookup of ResourceItems, which are already kept resident in IDEs for
example for layout rendering and inline resource folding. With this,
the lint rules can be do incremental analysis instead.

In addition to providing the machinery for looking up the resource
repository, this CL updates many key lint rules to run incrementally:

- The view cast detector, which checks that casts of findViewById
  calls match the types of the tags found in the layout XML files

- The string format check, which checks consistency between formatting
  call parameters and formatting string definitions

- The array size detector, which ensures that an array definition has
  the same number of children in other configurations

- The wrong id detector, which looks for @+id references in layout
  parameters and flags any references that are not bound to actual
  views anywhere

- The missing orientation detector (which now consults the styles
  referenced from the LinearLayout, if any). (Fixes issue #66249)

In addition, it also updates the IconDetector to include the xxhdpi
folder when checking that the dip sizes are roughly equivalent across
densities, and it bumps up the severity of duplicate id's found in a
single layout from warning to fatal. Also unit test infrastructure to
support checking incremental lint checks using a resource repository.
And finally it updates a couple of detectors (LocaleDetector,
ViewTagDetector) which had incorrect analysis scopes such that they
would not run incrementally even though they could.

Change-Id: I6a5fb3a38fb40b29d68df1bf30fa123c4515acbf