Package BuildConfig class in aar.
To do this we need to ensure that there are no conflicts between
the project and its dependencies. This means they all have different
package name.

Since we haven't enforced this before, there is a (temporary) way
to disable this enforcing. This also requires *not* packaging
BuildConfig in the aar.

DSL Options:

enforceUniquePackageName (app/lib): default true.
packageBuildConfig (lib): default true.

These options will be removed in 1.0 and will be always be true.

Change-Id: I553a4883b7f3580631af7d95ccaf722a7f70c95a