Add support for extracting support annotations in Gradle libraries
The new extract<Variant>Annotations task will look for support
library annotations in the sources, and create an annotations.zip
file containing XML descriptions of these annotations, which can
be digested by for example Android Studio's new ResourceType
inspection. For example, the library can define a new method like
this:
    public int draw(@DrawableRes int id) {
        ...

and then any clients of the library (when the AAR has been
built with annotations extracted) will get inspection warnings
if they attempt to pass for example an R.string id to this
method.

This task is automatically run as part of building a library AAR,
and it will do nothing if it does not find the android annotations
.jar as a direct dependency of the library.

Change-Id: I42f46dc1cebd7f120115ff897cb90436239c9ed1