Added validation step for android attributes.
Now all attributes which have an associated Validator in the model
will be validated after the xml is loaded and before it is attempted
to merge the file.

The model for all android's attributes will come in a subsequent CL,
this is the feature implementation.

Also added key uniqueness contraint so there cannot be 2 child of an
element of the same type with the same key (so far this is an error,
possibly downgradable to warning if the 2 elements are equals).

Change-Id: Iae48b6adb72684f196684c0a40323fa2bf5a8f5c