Add a Google Maps activity template.
This does not include any special support for generating an API key.
Instructions for getting an API key are displayed to the user after
going through the template wizard.

The API key is kept in a resource file to allow for easier merging.
For example, if the user adds their API key and generates another map
activity, their API key should not be overwritten.

Change-Id: If7edf693686d61a4a1c75686a4add61fd2c2108a