Added ability to save and reload manifest merger actions.
Split ActionRecorder which is not used to record actions during the
merging activities and Actions which is the resulting object once
merging is finished. Actions can be pesisted to an xml document
and reloaded from one.

Started provided some convenience APIs on Actions to retrieve actions
per element or attribute.

Implemented a simple ResultMapping to be able to get a pointer for
each resulting merged xml line to the original location that line
came from.

Change-Id: I8f073d34471928151cf1f12a84137390feb377ae