Initial version of implied manifest elements implementation.
When importing libraries targeting past versions of the SDK with some
implied permissions (and possible other element types) into an
application targeting a newer SDK, we must explicitely declare those
implied permissions for the application to run correctly.

Implemented a subset of the implied permissions, basically to hook up
the overall mechanism, the rest will be come in a subsequent CL.

A number of existing test needed modification as they declared old
SDK versions and triggered such permissions generation.

Change-Id: I87e196110390f775eadcaf620929ef402ad79441