WIP: jvm/host archive flags in SDK repository
The SDK repository is composed of "packages", each representing
one thing the user wants. Each package is actually composed of
one or more archives (zip files) that provide the actual files.
There can be more than one archive when we need to distribute
things that depend on the install host characteristics.

The current archive attributes are:
- os: one of any, windows, macosx, linux or ppc.
- arch: one of any, x86 or x86_64.

The change is:
- instead of being XML attributes, these become XML elements
  children of the parent archive element. It's easier to deal
  with elements than attributes in most XML APIs. The elements
  are optional -- if they are not present, it means they do not
  filter on this particular axis. Otherwise the host needs to
  match the filter requirement.
- "os" becomes "host-os". The ppc option is dropped as MacOS PPC
  is obsolete.
- "host-bits" represents the bit size (32-bits vs 64-bits) of the
  userland architecture.
- "jvm-bits" represents the bit size (32-bits vs 64-bits) of the
  JVM running the SDK tools. This matches the old "arch" attribute
  and removes confusion on what "arch" means.
- "min-jvm-version" is the minimal version of the JVM needed to
  use that package, e.g. "1.6" or "1.7".

WIP -- the CL only changes one of the schemas to get a feel for it
and does not properly use the attributes nor implement the filtering.
This is for review only --> DO NOT MERGE.

Change-Id: I49c9c2c02aa025c550437257b26b75a898e59b6a