Code Coverage support with Jacoco.
This uses the same mechanism as Gradle to dynamically depend on
some version of Jacoco.

Build Type has a new "instrumented" property allowing one to build an
instrumented apk. When testing an instrumented apk, the compiled code
gets an offline instrumentation with Jacoco and the Jacoco agent gets
packaged in. Post-test the coverage file gets extracted and a
report task creates an HTML report and an XML file for CI integration.

Change-Id: I5bd950326c920765adc96be8db38838c5e3703e3