Adds lint check for exporting PreferenceActivities.
Fragment injection gives anyone who can send a PreferenceActivity
an intent the ability to load any fragment, with any arguments,
into the activity's process. This is a security loophole. We fix
this by adding a lint check that checks that any PreferenceActivity
or subclass of PreferenceActivity is never implicitly or explicitly
exported in AndroidManifest.xml.

Change-Id: I487cdb8b8496a39cd4c915642826e4e88efd967d