Fix emulator console returning null every other call
The EmulatorConsole class maps open consoles to the emulator
based on the emulator's port number. This leads to issues in the following
scenario:
 - create mapping from port 5554 -> console 1
 - close emulator at port 5554
 - start new emulator which is now assigned port 5554
 - Attempt to getConsole():
    - this will see that there is already a mapping
    - however, pinging that console will fail
    - so null is returned

This problem occurs because the consoles are not cleaned when an
emulator gets killed. Since the emulator console itself has no knowledge
of when an emulator gets killed, this CL adds a new method that can
be called to close a particular emulator. This API can be called by
higher level code when it knows that the emulator corresponding to
a particular console has been killed, or pro-actively close connections
when not needed.

This possibly fixes https://code.google.com/p/android/issues/detail?id=68069.

Change-Id: I593e3b5e9ec452dfe608c783d1399ad34340deb5