Extends a11y check to handle views with onTouchListeners.
In accordance with GAR 1.4, views that use OnTouchListener
should keep their logic in an overridden view#performClick
and the OnTouchListener should call performClick on the
view. This lint checks that make by making sure that
any custom View that has setOnTouchListener called on it
also overrides performClick and that performClick is
called from any OnTouchListener's onTouch method.

Change-Id: Icf5adb6c329590fa7744cfa4ed11e8ddca73eb9a