sdklib: rename all XSD to a 2-digit number.
Since the sdk-repository xsd is in version 10, we need
all of these to have 2 digit numbers -- otherwise "10"
shows up before "1" and that makes complexifies the
sdk_repo.mk's logic. And it looks better anyway.

Note that in the XMLNS URI used in the xml files, we
still use numbers without a leading zero for backward
compatibility.

Change-Id: Ic0a90aa6afa6c679570a9a70f1dd7730a3c7dc7b