normalize qualifiers during resource merge.
For aapt -sw600dp is the same as -sw600dp-v13. This is because
that qualifier was introduced in API 13 and is ignored on lower
API.

This can trigger a problem if the merge is supposed to merge
res that are split across both -sw600dp and -sw60dp-v13. Aapt
will consider them to the be the version and (now) complain.

In order to fix this the merger will now "normalize" all res folders
based on the API level its qualifiers were introduced. This is stored
in the XML blob of the merger which is now in version 2. Version 1
will fail to load and force a full merge.

Change-Id: I3a1bdb894a30ff0e0b8864f107e378f58cc3434b