Update lint plurals check to ICU 53
This CL updates the plurals data used by the lint plurals check to be
based on ICU 53 instead of ICU 52, which among other things fixes this
bug:

75799: Incorrect lint check for plurals in Russian

However, more importantly it switches the plurals database
implementation over to be table driven (rather than manually encoded
as before), where the table is generated by code in the unit test that
is derived directly from plurals.txt.  This allows the database to be
much more easily and accurately updated in the future, and the CL also
uses this to extract sample data which is now used by the plural
detector to show specific examples of multiple numbers that match a
single quantity category.

The table approach should also be much more efficient; we don't have
to initialize large maps up front since in the vast majority of cases
programs will not be referencing it.

Change-Id: If7d8533baeefbfc0a5a8bbcbaaf4f3cfd3eb39bc