added a ManifestInjector to just do system property injection for libraries.
We used to invoke the full featured manifest merger when building flavored
lib which merged the transitive set of dependencies inside the resulting
library manifest file.

This could create issues down the road so instead, we now just invoke a
simple subset of the manifest merger capabilities to do placeholder
substitution and system property injection.

Change-Id: I3724b31ad18c16af475ee97a7dde5ea5dc78c35a