Add automatic resource shrinking to the Gradle plugin
This CL adds a new step to the build which for build types with
ProGuard shrinking enabled, will also strip out all the unused
resources - bitmaps, strings, statelists, arrays, etc.

This has two big advantags:

- The user doesn't have to do anything (such as locating and removing
  unused resources in the various folders); you get all the shrinking
  benefits automatically. This means it also figures out which
  resources don't apply in the current variant and strips them out, so
  you don't need to be as careful with for example placing resources
  in variant-specific source sets.

- It works with library projects. This means you can use library
  projects with a large number of resources (e.g. appcompat, play
  services), and only the resources that are actually needed by your
  application, transitively, will be included. The savings here
  are significant:
  - For one project which used appcompat and play services, the
    compressed binary XML resources shrunk from 1226KB to 767KB
  - For a default Studio project which uses appcompat for action bar,
    menus etc, the binary resources shrink from 1.5MB to 1.1MB
    (compressed size 389KB to 289KB). If you don't actually use
    any appcompat features that require resources, the resources
    will shrink down to 132K (which are basically our own launcher
    icons in the default project).
  - For Play Services, if you don't use features that require
    resources, it will shrink from 708K to 72K.

The resource stripper uses the ProGuard task's output, and the
manifest merger's output, to figure out the root resource references
needed by the app. It then computes which resources are "reachable"
from this set of roots (for example, a layout include tag will make
the included layout reachable, a style tag will make the corresponding
style reachable, its parent reference will then include the parent
style, and so on for state lists, attribute resource URLs, etc.)  It
will then remove all the unreachable resources: for file resources,
the resource files are removed, and for value resources the value.xml
files are rewritten without the unreachable resources.

Limitations:
- The task is currently not incremental; it deletes and replaces
  resources in the resource merger output directory. This needs to be
  fixed.
- It does not currently strip out assets. This could potentially lead
  to large savings.
- It currently does not strip out declare-styleable attributes. This
  will require some additional handling, but the potential savings are
  very small.
- There are some additional opportunities we can pursue, such as
  looking at the effective app minSdkVersion and removing all assets
  that are masked by any -vNN folder where NN >= minSdkVersion.

This CL is a work in progress; the way the task is hooked up is hacky,
we should support incremental execution, and error handling (for XML
parse and I/O errors) should be tightened up.

Change-Id: I3aac1a97034faf2bbeb47020ab9442a303298abf