Have the SDK download the Linux versions of the packages when running on FreeBSD.
The Android SDK is not officially supported on FreeBSD. However, since it mainly runs using Java, and since FreeBSD has a Linux compatibility layer, it does works as usual.
The only problem is that it doesn't recognize its host platform correctly, and thus downloads packages for Windows.
While FreeBSD can run the Linux version of the build-tools (aidl, aapt, etc.) and so on, it certainly can't run .exe files.
This patch fixes that by setting the right constants to their Linux value when the system uname is "FreeBSD".

Change-Id: Ib552ef363a7263bd4059e9b19f056a5b8c04f179