Fixes "The command line is too long" error
On Windows systems the command line can only accept commands that are 8192
characters or less.  Projects with many dependencies can sometimes exceed this
limit and the dex command will fail.  Right now the libraries are getting
passed in one at a time to dex.  Instead of doing that this patch copies all of
the libraries into a single directory and passes that directory to dex. This
keeps the command line short so dex will not fail on Windows systems.
Signed-off-by: John <john@john-VirtualBox.(none)>

Change-Id: I08efd0748cc744d234512c80ee6973210978f2e1