Selectively allow importing libraries with incompatible uses-sdk
when a project A(min sdk 8) that uses lib B(min sdk 14) only when the api is available and falls back if its not there,
we used to generate an error since the project's min sdk library was smaller than the library.

With this change, users will be able to override this mechanism by annotating the uses-sdk on the
project's AndroindManifest.xml with a tools:node="remove" and tools:selector="com.foo.lib1".
The presence of the selector is mandatory and checked to make sure we don't blindly override.

This will remove the uses-sdk from the imported library from the merging activities.

Change-Id: I1bc1bc974074d6771be65bc0852b093a9c0ca551

Conflicts:
	build-system/manifest-merger/src/test/java/com/android/manifmerger/ManifestMerger2Test.java