Switched tools: decorations removal to an optional feature.
We used to make decisions on wether or not tools: decorations should be
removed from the resulting merged document based on wether we are merging
an application or library. Now an explicit optional feature will drive the
removal. This will be useful to Studio who will use the merger to integrate
templates and never want to remove the tools: decorations wether it is a
library or an application

Change-Id: I56690179408eba14b090bb00282747b80b4ce716