Fixes command line too long error on Windows.
This depends on change 103691 which added the
--input-list option to dex. This change writes all of
the library jars to a file and then passes the filename
to the --input-list option in dex.  This allows us to
keep the command line short and keeps Windows from
complaining.  We also check to see if the build tools
revision is greater than or equal to 21.  If not we
revert to the old way of doing things.

Change-Id: If59341ea4663af0c5de73535c69d1720824768f9
Signed-off-by: John Fazekas <jfazekas@fd-software.com>
(cherry picked from commit 9015842c7cca1d6001cb4c00c61422d509ce1075)