Add API to get multiple matching resources.
If a FolderConfiguration doesn't specify a Qualifier, all resources
specifying that Qualifier are accepted. This change adds an API to get
all matching resources in case it's not possible to reduce the number of
matching resources based on the available qualifiers to one.

Also, for a Qualifier missing in the FolderConfiguration, if resource A
didn't specify any qualifer and resource B specified one, we now
consider 'A' to be a matching resource. This is necessary so that the
ConfigurationMatcher can unset some Qualifiers and get all resources
that differ only based on that Qualifier. This is important for ignoring
the API levels in the layouts.

Bug: http://b.android.com/75156
Change-Id: I4ad8559a111ea6cfd49b8783d4af503fd2bd8a34