72830: Editing XML in additional res directories is missing functionality
The code which adds a sync listener (to update the set of known
resource directories used for indexing, layout resource lookup etc)
was in the wrong place; it was only done when we encounter a gradle
project but don't actually have a loaded model yet. This used to
always happen briefly, but no longer does now that we usually cache
models across IDE invocations. The right fix is to add this listener
unconditionally for both the have-model and don't-have-model cases.

Change-Id: I8610282d2f17161cf977dea80d045d905c8b5073