Don't query new devices immediately for system properties
This CL removes the code that immediately queries a connected device
for its system properties. It turns out that by the time this query
is given, the device is usually not in a state to respond to the
query (it is usually in the unauthorized state at this point, most
likely waiting for the security key to be validated).

Change-Id: Idea5ac6d6df488ea2ff85412204a75afe9824751