Support the ANDROID_SERIAL property for targeting devices
Add a check in the ConnectedDeviceProvider for a specific
device serial to be targeted against based off the value
stored in the ANDROID_SERIAL environment variable.

The default behavior of running across all connected devices
will occur if this variable is not defined or is empty. An
exception will be thrown if the targeted DSN is not found.

Defect: 75407

Change-Id: Ib4dba8398cbb249420d13eb48da076ac67c61609