Clean up ProductFlavor/BuildType API.
WIP.

1. Enforce Immutability on some items returned in by the variant APIs
(variant collection and variantFilter)

ProductFlavors and BuildTypes returned by a variant through the variant
API should be considered read-only. If there are not there are some
potential issues:
- the change would be global to the build type, impacting all variants
  using the build type. This is potentially confusing to developers
  who would (most likely based on seeing new users reacting to gradle)
  think it only affects the current variant.
- the changes may actually not necessarily impact the tasks depending
  on how the tasks are setup, or if task creation is based on the
  property being changed (since this is too late to impact task
  creation).

2. Move NdkConfig out of the base Flavor/BuildType classes.
This is not ready and shouldn't be included in the builder library.

3. Make signing config part of the Flavor/BuildType based interfaces.

4. Fix ProductFlavor API by having it never returning primitive types.
This makes it impossible to distinguish between no value set and default
values which is critical in ensuring that flavors properly merge on top
of each other (include the default Config).

Part of these fixes, the VariantConfiguration class is now using
generics to help dealing with custom implementations of ProductFlavor
and BuildType in the Gradle plugin.

Bug: https://code.google.com/p/android/issues/detail?id=76313
Change-Id: If7e069e6398f59a6f274fd113a4599a8a04f42b8