PreDexing support for 21+ multi-dexing.
Change the pre-dex caching to handle 1+ output
files.

When multi-dex and pre-dexing is enabled, the dex
task doesn't receive the dependencies. Instead
they are directly added to the apk as extra dex.

When copying the dex files into the APK, we rename
them classes<N>.dex, starting with 1 since the platform
expects continuous numbers, and stop on the first missing
one.

Change-Id: I016aca148992eb42a73b569e91ff45ae8b22d89a