Fix legacy mode for multidex.
Dex files were added to the apk with a renaming
mode to ensure we have continuous classes<N>.dex names.

However this only worked in legacy mode if the list of dex
files returned by File.listFiles had the main dex file first
in the list. Otherwise anothe dex files would be put as the
main dex file in the APK.

Change-Id: I0d7459932f0c832a8a4ed9dc426cad4a0099b682