Add support for variant-specific dependency.
This is only active if there is a flavor (since
build type == variant without flavors).

In case of multiple flavor dimensions there is only a
"multi-flavor".

The priority order is now:

Variant-specific, build type, multi-flavor, flavor(1..N), default

Limitation: Variant-Specific and Multi-Flavor (gradle) configuration
objects are created by the plugin late in the evaluation (actually
in the postEvaluation step), making it impossible to directly using them
the build file. This is mostly a timing issue with no easy work around
until the new model mechanism.

To work around this, one can simply manually create them ahead of time:

configurations {
  fooDebugCompile
}

android {
  productFlavors {
    foo { ... }
  }
}

dependencies {
  fooDebugCompile '...'
}

When the plugin creates the variant and attempt to create the variant
specific configuration, it'll see one is already there and use it
instead.

Change-Id: I1e24ac27efd038342b500ac9800ce3af56df9673