Add sources for generate annotation task.
registerJavaGeneratingTask() now adds the output directory to the
source of the generate annotation task if it exists.

When that method is called, the sourceset of the generate annotation
task has already been populated, so it needs to be explicitly added.

Fixes https://code.google.com/p/android/issues/detail?id=76139

Change-Id: Ieb4dd2e67f941aaff57260c01783b012a2f56a5b