Apply Android Studio branding to tip of the day.
And a few other places.

Change-Id: Id50e4fefd9a30ad6dd11690b4ec28ca0fe3d7a82

Rename lintCompile field to reflect the real task name.

What is supposed to live in src/lint/java? Do we still wan to support it?

Change-Id: I3c81d285e8bc4b3ac5e25862068cdb8041e98713

Rename lintCompile field to reflect the real task name.

What is supposed to live in src/lint/java? Do we still wan to support it?

Change-Id: I3c81d285e8bc4b3ac5e25862068cdb8041e98713