Add an option to override the Gradle version check.
The java property is com.android.build.gradle.overrideVersionCheck.
You can set the override to true using
-Dcom.android.build.gradle.overrideVersionCheck=true

If true, the check will still happen, but a failure will
become a warning, not an exception like normal.

This can be useful for testing the Android plugin with newer (or older)
Gradle versions.

Change-Id: I9d75a2f69b29cd426806c5e123280b19dae14fc6