Checking whether emulator is actually running
The emulator creates a lock file containing the pid of the
emulator. Studio declines to launch the emulator if the lock
exists but doesn't actually check to see if the referenced
process exists, so stale lock files are an issue.
Adding a check to see if the process exists.
Fixes https://code.google.com/p/android/issues/detail?id=81860
in that obsolete files shouldn't be an issue after this change.

Change-Id: Ib2a5324f9853dcd52a4f8b235084c4f6598abd49