New process executors.
The goal of this is to replace CommandLineRunner in way
that's easier to use and replaceable with Gradle's way
of running processes.

Gradle has a special API for running Java-based external
processes so we have 2 executors to match Gradle's API.

The underlying motivation is making sure external Java
based processes (dx/Jack/Jill) run with the same Java
that Gradle is using.

Change-Id: I4163e94346c1aa99f41f41d096f33526cd523f97