Add typed constructor support to lint JavaScanners
The JavaScanner interface in lint allows detectors to specify method names
they're interested in and a special callback is invoked for just those
methods rather than having to provide their own visitor.

This CL adds the same things for constructors: detectors can specify
types they are interested in, and they will then be called for each
constructor invocation of the given type.

This CL also rewrites the DateFormatDetector to use this mechanism.

Change-Id: I1239e949ef9cde50568633981f3a0f151db997b5