Sync test and tested dependency trees.
When Android loads a test app, both apps are in the same
classpath. This means that both apk cannot contain the same
classes. This can easily happen when both the app and the test
app has some identical dependencies.

This changeset includes a check that detects possible cases
like this and attempts to resolve them or error out.

- If the dependencies are exactly the same, then skip the
  packaging of the classes in the test app.
- If the dependencies have different versions, break out.

We can improve #2 later by checking if the app dependency is more
recent than the test app, then we could upgrade the test dependency
to the same version, but that can have side effects as well.

Change-Id: I472d0fd9800da3ca46832f42d084ef96f183924f