Fix @+id resolution in some cases.
If an id name in the project namespace is created using @+id notation,
and if one with the same name exists in the android namespace, the
resource resolver returned the framework id instead of creating a new
one. This makes the resource resolver return null, indicating that the
id is not a reference and the new id is then created for it when
queried for it.

This change also skips a warning which is logged when referencing a
style attribute that is not defined in the current configuration. For
example, if a widget defines its style as "style=?attr/foo" and foo is
not defined in the current theme. This is because even though it's an
error in most cases, the framework doesn't throw an error. And more
importantly, the support library (appcompat) does this. Showing a
warning in this case will confuse more users that it will help.

Change-Id: I89db6cc05014969ad38c8870544b68be352c2a7c