Added abi specific test and removed all jniFolders when pure splits are enabled
While working on task parallelization, I realized the combined abi and density
pure split test was not running.
After adding it, I realized that although the splits were correct, the main
APK was still containing all the .so files.

The fix is a bit sensitive, to be discussed further during code review,
but basically, when there are abi pure splits, we do not provide
any jniFolders to the main APK package task as all jni resources
are packaged in their respective split apk.

also added a missing reset() the build.gradle for the test.

Change-Id: Ie1d9d910db796ca943caeea5a307f1dd6acda005