Remove unnecessary use of convention mapping.
Convention mapping are used to avoid configurations if the tasks are
not executed, but unnecessary if the configurations are trivial.

This remove some of the use of convention mapping.  In particular, it
also remove the use of project.file when used with project.buildDir.
project.buildDir always returns an absolute path, so project.file is
simply creating a new File.

Change-Id: I55016a5865053512e1f32037684983682e13755a