Move LegacyCallback to layoutlib-api.
This move effectively changes the API from interface to abstract class.
This enables adding methods with a default implementation possible
without having to increment the API version number. However, note that
the methods added without incrementing the API version should not be
called from the LayoutLib. This prevents accidently calling a method
that simply doesn't exist on the IDE side.

LegacyCallback in sdk-common is not removed in this commit, to prevent
build breaks.

Change-Id: I31f77c1bc05fcea6ebaffba7a7e160e455bddfe2