Move LegacyCallback to layoutlib-api.
This move effectively changes the API from interface to abstract class.
This enables adding methods with a default implementation possible
without having to increment the API version number.

Care must be taken to call such newly added methods from LayoutLib, as
they may not exist on older IDEs. Either the API version should be
incremented to prevent the LayoutLib from being loaded on older IDEs, or
have the calls in LayoutLib guarded by a check to ensure that the method
will exist. Future change creates an API which can be used to implement
such a guard (Change-Id: Ibd1c32bd2edbc15f8ae98e9508637df28bda0906).

LegacyCallback in sdk-common is not removed in this commit, to prevent
build breaks.

Change-Id: I31f77c1bc05fcea6ebaffba7a7e160e455bddfe2