Namespace aware DeclareStyleableResourceValue.
Update DeclareStyleableResourceValue to be namespace aware. Thus if a
declare-styleable resource references two attrs with the same name in
difference namespaces, both will co-exist in the list of attrs.

This changes the methods in layoutlib-api, which ideally should never
be changed. But in this case, the methods that are updated were never
used by any version of the LayoutLib and hence should be okay.

Change-Id: I6950289e5a196e4e2fd58cafe82d6fa3b3a5d2a5