Allowing a MajorRevision to be parsed from a FullRevision
This is an interim change.
To simplify the code, RemoteSdk will write a FullRevision into
Source.Properties in all cases.
Next I will try to switch package revisions to be
PreciseRevisions in each case, which should allow us to preserve
backward compatibility (without the need for this change) and simplify
the code even further.

Change-Id: Ie316f34fa1f72e513bb9a0e798891c27d447efc7