Fix dependencies on tested component for lib test.
With VariantDependencies implementing ConfigurationProvider,
it very easy to think that one can pass a VariantDependencies
in the list of ConfigurationProvider when computing another
VariantDependencies.

However this breaks the provided scope for tests. This was
not apparent before due to aar not being supported as provided.

The issue is due to the composite configurations compile/package
being made of provided/compile and compile/package respectively.

When passing a VariantDependencies(1) to the compute method, to
create a new VariantDependencies(2), one would end up with:

provided(1) -> null
compile(1) -> provided(0) + compile(0)
package(1) -> compile(0) + package(0)

provided(2) -> null
compile(2) -> provided(1)+ compile(1)
package(2) -> compile(1) + package(1)

Unraveling the transitive dependencies, we end up with
package(2) -> provided(0), compile(0), package(0)

This shows that provided(0) ends up in package(2) which is not correct.

The solution is to handle ConfigurationProvider and VariantDependencies
as different types of input in compute()

Change-Id: Ia1b77304440976724446c5f9b75d5519b0163ba3