Transitive deps for separate test modules.
The classes for the tested app is coming through
a classes.jar artifact which brings, transitively
all the app dependencies.

This means that we get into a situation where we get jars
that depends on aars which is not a typical situation we
want to support.

This changeset detects the particular case of a separate
test module and handles it by moving the aar at the top-level
and marking them as Optional to make sure that they don't
get packaged.

This works except for the transitive dependencies showing
up in the test app directly and through the tested module.

Change-Id: Ic4f47709b1cf695021bb462f64aa3c93874a74aa