Persist merge blame log.
It can then be used to rewrite AAPT error messages.

This CL will break MessageRewriteTest, as the logic that inserts source
comments is removed. It is fixed by "Rewrite AAPT messages using new
persisted log." It also fixes a potential NPE if a ConsumerException is
caused by an exception without a message.

Change-Id: I2756068189226eed08d96f64e2ca1b330e5d2dd7