Allow duplicates in public.xml.
Public resource items have types just as other resource items, but they
were all being considered as ResourceType.PUBLIC, when actually they
could have colliding names.

For example, there could be an ID resource and a string with the same
name.

This CL changes the internal string key to include the subtype of the
public resource in its name, and relaxes the duplicate check for public
resources.

https://code.google.com/p/android/issues/detail?id=170153

Change-Id: I88b42fb9c777210b7e976ee883baf16a37a26ae6