Fix issue with specify multiple C/C++ flags.
Gradle accepts an array of args for c/c++ flags.  Each flag must be a single
element in the array.  Therefore, we need to modify the flags in NdkConfig
to be a list.

Getting rid of the ManagedSet<ManagedString> until Gradle provides a better
method for defining a Managed list.  This improves the DSL slightly.

Change-Id: I5cfcb414c84b6ddcd09a48d05f691b3bc9660976