rework of the way we process java resources.
Instead of processing java resources during the packaging of the APK,
moved this upfront before the obfuscation tasks. This will allow
the obfuscation tasks to have a chance to adapt the java resources
following packages obfuscation.

wired up both Proguard and Jack to use the new merged java resources
folders.

Change-Id: I23cfd5be5e62c62eb2759e53692984642856a55a