Don't associate stack traces with a specific heap.
Previously stack traces were associated with the current heap at the
time the stack traces were loaded from the hprof file. This caused
stack traces not to be found if the current heap had changed since
loading the stack trace.

This commit stores all stack traces in the snapshot instead of
associating them with heaps.

Change-Id: Ie4b107f3536037a7d4d29c820ec69f067fce92d9