ddmlib: Remove locking around mDevices in DeviceMonitor
This CL removes all locking around an internal field (mDevices)
inside the DeviceMonitor by replacing it with a CopyOnWriteArrayList.

This list is only updated when the number of devices changes, and
the number of devices itself is pretty low. So there is really no
overhead to using a CopyOnWriteList.

Additionally, the existing code had to synchronize its access with
2 locks: first from the debug bridge, then the lock on the field
in order to always lock them in the same order. Now that no
locking is necessary, the debug bridge's lock can be made private
to the bridge.

Fix for https://code.google.com/p/android/issues/detail?id=172387

Change-Id: Idd9089bf09edd33135c9f17e1b76babe1e3d62e8