If attempting to use multiple resource sets that are nested, the startsWith() logic ends up picking the 1st folder instead of the 2nd.
One of the implications is that file changes detected will not work.   Even though incremental task runs detect that the file has changed, the file
does not appear to get updated as a result when using nested resource sets.

This change makes the logic avoid using startsWith() and replaces it with checking the pathnames are the same.

Change-Id: I48db691f78f2178b938791c03e488b00e0a58964