Make Jacoco instrumentation incremental.
This replaces the task that ran the jacoco ant task.

It has the disadvantage of tying the jacoco version to the gradle plugin
version, as it is now a direct dependency. However, users can upgrade jacoco by
adding a newer version as a build script dependency.

If we want to completly decouple the versions, a separate android jacoco plugin
seems like the best solution.

Change-Id: Icbb4ece147a9f25c50e2c4bc26e36b1f5bea84da