Change the new activity templates to generate an App Bar.
So far: only BlankActivity has been addressed.

For a new project and a project already with appCompat:
Show a button (checked by default) for "Include App Bar".
For existing projects without appCompat this button is hidden.
If checked the layout will include several Material Design
controls. The AppTheme will be Theme.AppCompat.Light.NoActionBar
with some color overrides.
The design and appCompat will be added to the gradle dependencies
for all new projects. The design library will be added to
projects that already has appCompat.

Change-Id: I107679b6b1953d13eda428d6f50a53c934140642