Update BlankActivity template to add some options
The BlankActivity template is very useful, but is by default setup for
beginners to android - it not only creates a blank activity and adds it
to the manifest etc, but it also generates a default menu and adds a
'hello world' string and TextView to the layout. Whilst this is great
for beginners, it's annoying to have to continually delete these
same things over and over.

This commit adds options for these two things, by adding checkboxes to
the template UI - defaulting to true to not change the current
behaviour, the user can now choose whether or not to add the
menu/Hello World components

Change-Id: Ib767effbfe468e2aafe191c89453e7d09f4adb51