Transform API.
During setup of the tasks, we often have cases
where optional tasks consume and transform elements.

This makes it difficult for the downstream tasks to
be properly configured as they have to handle all
possible combination upstream. For instance having
or not having jacoco, proguard, multi-dex, etc...
complicate things.

The goal of the transform API is to automate this
for a subset of the types handled by the build
(classes, dex, java resources).

A Transform indicate which type and which scope it
applies to and how it transforms them. The badly-named
TransformPipeline handles adding transforms, creating
intermediary streams between transform and setting up
dependencies.

This is a first pass commit creating the basic API, and
a couple of Transform (though not using them yet).
The API is not final and will slightly evolved in the
next few commits as things stabilize.


Change-Id: Iefe59a60abf0fe946a32ae24747993a072bcb80e