Check for test run failures before reporting a successful run
If a test run fails in a BeforeClass or AfterClass method (any place that isn't actually part of a test), test execution will stop and the error will be reported as a test run failure and there will be zero failed tests. This can lead to scenarios where some subset of the tests run, and then a test run error is encountered which causes the rest of the tests to not be executed, but the run will still be reported as successful since we were only checking for failed tests, not failed test runs.

Change-Id: If947fc62505ee4eafc76e883e71cd629547e6827
Signed-off-by: Drew Hannay <drewhannay@gmail.com>