Handle Java Resources through Transforms.
For this to work we need to properly handle streams that
represents both class and resources but only get consumed
for one. This will create a copy stream with the remaining
types while the consumed type is transformed into a new
stream.

Also add more fine grained streams for the classes, notably
around dependencies.

Finally add some tests for the Transform manager.

Change-Id: I8973c1e157f27662db2e9be58811227f5bb53415