Add lint check for insecure TLS TrustManager
This lint check detects declarations of TLS TrustManagers that do
not properly check certificates. The check looks for empty
checkClientTrusted and checkServerTrusted methods as well as
getAcceptedIssuers methods that always return null or an empty
array. We do not check if or how the declared TrustManager
is actually used by the application.

As documented by numerous sources such as [1] [2] [3], Android
applications commonly fail to properly check certificates while
establishing TLS sessions, making them susceptible to MITM attacks.

[1] Fahl, et al. Why Eve and Mallory Love Android: An Analysis of Android SSL (In)Security.
CCS '12.

[2] Sounthiraraj, et al. Large Scale, Automated Detection of SSL/TLS
Man-in-the-Middle Vulnerabilities in Android Apps. NDSS '14.

[3] FireEye. SSL Vulnerabilities: Who listens when Android applications talk?
https://www.fireeye.com/blog/threat-research/2014/08/ssl-vulnerabilities-who-listens-when-android-applications-talk.html

Change-Id: Id0bb4304a000f7ca36fd9433659b63dd2a49256b