Add lint checks for insecure broadcast receivers
We propose two new lint checks to detect insecure broadcast receivers that
fail to properly check the origin of received intents, potentially making
the receiver vulnerable to spoofed intents.

The first check is for broadcast receivers that have declared an intent-filter
for a protected-broadcast action string but fail to actually check the received
action string, as described by section 4.3 of [1].

The second check is for broadcast receivers that have declared an intent-filter
for the SMS_DELIVER or SMS_RECEIVED action string but fail to ensure
that the sender holds the BROADCAST_SMS permission. [2] [3]

Note that neither of these checks address receivers dynamically created at runtime.
They only address receivers that are declared in the application manifest.

[1] Chin, et al. Analyzing Inter-Application Communication in Android.
Mobisys '11. https://www.eecs.berkeley.edu/~daw/papers/intents-mobisys11.pdf

[2] https://commonsware.com/blog/2013/10/06/secured-broadcasts-sms-clients.html

[3] http://android-developers.blogspot.com/2013/10/getting-your-sms-apps-ready-for-kitkat.html

Change-Id: I2d1a20b2572b332a76d2fbf6cb5b0ff83cc7fab3